+++ packages/contracts/contracts/zkopru/controllers/UserInteractable.sol
@@ -1,5 +1,5 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
-pragma solidity = 0.7.4;
+pragma solidity =0.7.4;
 
 import { IERC20 } from "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import { IERC721 } from "@openzeppelin/contracts/token/ERC721/IERC721.sol";
@@ -10,7 +10,8 @@ import { Hash, Poseidon3, Poseidon4 } from "../libraries/Hash.sol";
 import { WithdrawalTree, Blockchain, Types } from "../libraries/Types.sol";
 
 contract UserInteractable is Storage {
-    uint256 public constant SNARK_FIELD = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 public constant SNARK_FIELD =
+        21888242871839275222246405745257275088548364400416034343698204186575808495617;
     uint256 public constant RANGE_LIMIT = SNARK_FIELD >> 32;
     using MerkleTreeLib for *;
     using SafeMath for uint256;
@@ -38,7 +39,7 @@ contract UserInteractable is Storage {
     ) public payable {
         _deposit(spendingPubKey, salt, eth, token, amount, nft, fee);
     }
-    
+
     /**
      * @notice Users can withdraw notes when only after they're finazlied.
      * @param note Note hash in layer 2. It is a poseidon hash
@@ -64,18 +65,19 @@ contract UserInteractable is Storage {
         uint256 leafIndex,
         uint256[] memory siblings
     ) public {
-        return _withdraw(
-            note,
-            owner,
-            eth,
-            token,
-            amount,
-            nft,
-            callerFee,
-            blockHash,
-            leafIndex,
-            siblings
-        );
+        return
+            _withdraw(
+                note,
+                owner,
+                eth,
+                token,
+                amount,
+                nft,
+                callerFee,
+                blockHash,
+                leafIndex,
+                siblings
+            );
     }
 
     /**
@@ -101,44 +103,34 @@ contract UserInteractable is Storage {
         uint256 prepayFeeInToken,
         bytes memory signature
     ) public payable {
-        bytes32 withdrawalHash = _withdrawalHash(
-            note,
-            owner,
-            eth,
-            token,
-            amount,
-            nft,
-            callerFee
-        );
+        bytes32 withdrawalHash =
+            _withdrawalHash(note, owner, eth, token, amount, nft, callerFee);
         require(!Storage.chain.withdrawn[withdrawalHash], "Already withdrawn");
 
         address newOwner = Storage.chain.newWithdrawalOwner[withdrawalHash];
         address currentOwner = newOwner == address(0) ? owner : newOwner;
         address prepayer = msg.sender;
-        bytes32 payInAdvanceMsg = keccak256(
-            abi.encodePacked(
-                prepayer,
-                withdrawalHash,
-                prepayFeeInEth,
-                prepayFeeInToken
-            )
-        );
+        bytes32 payInAdvanceMsg =
+            keccak256(
+                abi.encodePacked(
+                    prepayer,
+                    withdrawalHash,
+                    prepayFeeInEth,
+                    prepayFeeInToken
+                )
+            );
         // verify original owner's signature
         require(
-            _verifySignature(
-                currentOwner,
-                payInAdvanceMsg,
-                signature
-            ),
+            _verifySignature(currentOwner, payInAdvanceMsg, signature),
             "Invalid owner signature"
         );
-        uint ethToWithdraw = eth.sub(prepayFeeInEth);
-        uint tokenToWithdraw = amount.sub(prepayFeeInToken);
-        require(msg.value == eth, 'not enough ether');
+        uint256 ethToWithdraw = eth.sub(prepayFeeInEth);
+        uint256 tokenToWithdraw = amount.sub(prepayFeeInToken);
+        require(msg.value == eth, "not enough ether");
         // prepay tokens
         if (Storage.chain.registeredERC20s[token]) {
             IERC20(token).transferFrom(prepayer, currentOwner, tokenToWithdraw);
-        } else if (Storage.chain.registeredERC721s[token]){
+        } else if (Storage.chain.registeredERC721s[token]) {
             revert("Does not support NFT prepay");
         }
         // prepay ether
@@ -157,13 +149,25 @@ contract UserInteractable is Storage {
         uint256 fee
     ) internal {
         // range check
-        require(amount < RANGE_LIMIT, "Too big value can cause the overflow inside the SNARK");
-        require(eth < RANGE_LIMIT, "Too big value can cause the overflow inside the SNARK");
-        require(fee < RANGE_LIMIT, "Too big value can cause the overflow inside the SNARK");
+        require(
+            amount < RANGE_LIMIT,
+            "Too big value can cause the overflow inside the SNARK"
+        );
+        require(
+            eth < RANGE_LIMIT,
+            "Too big value can cause the overflow inside the SNARK"
+        );
+        require(
+            fee < RANGE_LIMIT,
+            "Too big value can cause the overflow inside the SNARK"
+        );
         require(nft < SNARK_FIELD, "Does not support too big nubmer of nft id");
         // check eth value
         require(eth.add(fee) == msg.value, "Inexact amount of eth");
-        require(Storage.chain.stagedSize < 1024, "Should wait until it is committed");
+        require(
+            Storage.chain.stagedSize < 1024,
+            "Should wait until it is committed"
+        );
         // check note fields
         require(_checkNoteFields(eth, token, amount, nft));
 
@@ -175,19 +179,25 @@ contract UserInteractable is Storage {
         uint256 note = Poseidon3.poseidon(resultHashInputs);
         // Receive token
         if (token != address(0) && amount != 0) {
-            try IERC20(token).transferFrom(msg.sender, address(this), amount) {
-            } catch {
+            try
+                IERC20(token).transferFrom(msg.sender, address(this), amount)
+            {} catch {
                 revert("Transfer ERC20 failed");
             }
         } else if (token != address(0)) {
-            try IERC721(token).transferFrom(msg.sender, address(this), nft) {
-            } catch {
+            try
+                IERC721(token).transferFrom(msg.sender, address(this), nft)
+            {} catch {
                 revert("Transfer NFT failed");
             }
         }
         // Update the mass deposit
-        Storage.chain.stagedDeposits.merged = keccak256(abi.encodePacked(Storage.chain.stagedDeposits.merged, note));
-        Storage.chain.stagedDeposits.fee = Storage.chain.stagedDeposits.fee.add(fee);
+        Storage.chain.stagedDeposits.merged = keccak256(
+            abi.encodePacked(Storage.chain.stagedDeposits.merged, note)
+        );
+        Storage.chain.stagedDeposits.fee = Storage.chain.stagedDeposits.fee.add(
+            fee
+        );
         Storage.chain.stagedSize = Storage.chain.stagedSize.add(1);
         // Emit event. Coordinator should subscribe this event.
         emit Deposit(Storage.chain.massDepositId, note, fee);
@@ -206,43 +216,47 @@ contract UserInteractable is Storage {
         uint256[] memory siblings
     ) internal {
         // range check
-        require(amount < RANGE_LIMIT, "Too big value can cause the overflow inside the SNARK");
-        require(eth < RANGE_LIMIT, "Too big value can cause the overflow inside the SNARK");
-        require(callerFee < RANGE_LIMIT, "Too big value can cause the overflow inside the SNARK");
+        require(
+            amount < RANGE_LIMIT,
+            "Too big value can cause the overflow inside the SNARK"
+        );
+        require(
+            eth < RANGE_LIMIT,
+            "Too big value can cause the overflow inside the SNARK"
+        );
+        require(
+            callerFee < RANGE_LIMIT,
+            "Too big value can cause the overflow inside the SNARK"
+        );
         require(nft < SNARK_FIELD, "Does not support too big nubmer of nft id");
         // check note fields
         require(_checkNoteFields(eth, token, amount, nft));
         // check the reference block is finalized
         require(Storage.chain.finalized[blockHash], "Not a finalized block");
         uint256 root = Storage.chain.withdrawalRootOf[blockHash];
-        bytes32 withdrawalHash = _withdrawalHash(
-            note,
-            owner,
-            eth,
-            token,
-            amount,
-            nft,
-            callerFee
-        );
+        bytes32 withdrawalHash =
+            _withdrawalHash(note, owner, eth, token, amount, nft, callerFee);
         // Should not allow double-withdrawing
         require(!Storage.chain.withdrawn[withdrawalHash], "Already withdrawn");
         // Mark as withdrawn
         Storage.chain.withdrawn[withdrawalHash] = true;
         // Check whether new owner exists
-        address to = Storage.chain.newWithdrawalOwner[withdrawalHash] != address(0)
-            ? Storage.chain.newWithdrawalOwner[withdrawalHash]
-            : owner;
+        address to =
+            Storage.chain.newWithdrawalOwner[withdrawalHash] != address(0)
+                ? Storage.chain.newWithdrawalOwner[withdrawalHash]
+                : owner;
 
         // inclusion proof
-        bool inclusion = Hash.keccak().merkleProof(
-            root,
-            uint256(withdrawalHash),
-            leafIndex,
-            siblings
-        );
+        bool inclusion =
+            Hash.keccak().merkleProof(
+                root,
+                uint256(withdrawalHash),
+                leafIndex,
+                siblings
+            );
         require(inclusion, "The given withdrawal note does not exist");
         // Withdraw ETH & get fee
-        if(to == msg.sender) {
+        if (to == msg.sender) {
             _sendEth(to, eth.add(callerFee));
         } else {
             _sendEth(to, eth);
@@ -251,8 +265,11 @@ contract UserInteractable is Storage {
         // Withdraw tokens if exists
         if (Storage.chain.registeredERC20s[token]) {
             IERC20(token).transfer(to, amount);
-        } else if (Storage.chain.registeredERC721s[token]){
-            require(nft != 0, "Circuit cannot accept NFT id 0. Please deposit other NFT.");
+        } else if (Storage.chain.registeredERC721s[token]) {
+            require(
+                nft != 0,
+                "Circuit cannot accept NFT id 0. Please deposit other NFT."
+            );
             IERC721(token).transferFrom(address(this), to, nft);
         }
     }
@@ -266,17 +283,18 @@ contract UserInteractable is Storage {
         uint256 nft,
         uint256 callerFee
     ) internal pure returns (bytes32) {
-        return keccak256(
-            abi.encodePacked(
-                note,
-                owner,
-                eth,
-                token,
-                amount,
-                nft,
-                callerFee
-            )
-        );
+        return
+            keccak256(
+                abi.encodePacked(
+                    note,
+                    owner,
+                    eth,
+                    token,
+                    amount,
+                    nft,
+                    callerFee
+                )
+            );
     }
 
     function _verifySignature(
@@ -296,7 +314,10 @@ contract UserInteractable is Storage {
             // final byte (first byte of the next 32 bytes).
             v := byte(0, mload(add(sig, 96)))
         }
-        bytes32 prefixedHash = keccak256(abi.encodePacked("\x19Ethereum Signed Message:\n32", message));
+        bytes32 prefixedHash =
+            keccak256(
+                abi.encodePacked("\x19Ethereum Signed Message:\n32", message)
+            );
         return signer == ecrecover(prefixedHash, v, r, s);
     }
 
@@ -307,24 +328,33 @@ contract UserInteractable is Storage {
         uint256 nft
     ) internal view returns (bool) {
         if (token == address(0)) {
-            require(nft == 0 && amount == 0, "Ether note does not have amount field & nft field");
+            require(
+                nft == 0 && amount == 0,
+                "Ether note does not have amount field & nft field"
+            );
             require(eth != 0, "Should have ETH field");
         } else {
             // this note contains token value
             bool isERC20 = Storage.chain.registeredERC20s[token];
             bool isERC721 = Storage.chain.registeredERC721s[token];
-            require(isERC20 || isERC721, "Not a registered token. Reigster that token first");
+            require(
+                isERC20 || isERC721,
+                "Not a registered token. Reigster that token first"
+            );
             if (isERC20) {
                 require(nft == 0, "ERC20 does have NFT field");
-            } else if (isERC721){
-                require(nft != 0, "Circuit cannot accept NFT id 0. Please deposit other NFT.");
+            } else if (isERC721) {
+                require(
+                    nft != 0,
+                    "Circuit cannot accept NFT id 0. Please deposit other NFT."
+                );
                 require(amount == 0, "ERC721 does have amount field");
             }
         }
         return true;
     }
 
-    function _sendEth(address to, uint val) internal {
+    function _sendEth(address to, uint256 val) internal {
         if (val > 0) {
             (bool success, ) = to.call{ value: val }("");
             require(success, "Failed to send ETH");
